single tx reduce fee-collecotr's access
