+++ libs/cosmos-sdk/x/auth/refund/refund.go
@@ -7,14 +7,14 @@ import (
 )
 
 func RefundFees(supplyKeeper exported.SupplyKeeper, ctx sdk.Context, acc sdk.AccAddress, refundFees sdk.Coins) error {
-	coins := supplyKeeper.GetFeeFromBlockPool()
+	coins := supplyKeeper.GetFee()
 
 	if !refundFees.IsValid() {
 		return sdkerrors.Wrapf(sdkerrors.ErrInsufficientFee, "invalid refund fee amount: %s", refundFees)
 	}
 
 	// verify the account has enough funds to pay for fees
-	_, hasNeg := coins.SafeSub(refundFees)
+	fee, hasNeg := coins.SafeSub(refundFees)
 	if hasNeg {
 		return sdkerrors.Wrapf(sdkerrors.ErrInsufficientFunds,
 			"insufficient funds to refund for fees; %s < %s", coins, refundFees)
@@ -24,6 +24,11 @@ func RefundFees(supplyKeeper exported.SupplyKeeper, ctx sdk.Context, acc sdk.Acc
 	if err != nil {
 		return sdkerrors.Wrapf(sdkerrors.ErrInsufficientFunds, err.Error())
 	}
+	err = supplyKeeper.AddCoinsToFeeCollector(ctx, fee)
+	if err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInsufficientFunds, err.Error())
+	}
+	supplyKeeper.ResetFee()
 
 	return nil
 }
