+++ libs/cosmos-sdk/x/supply/internal/keeper/keeper.go
@@ -2,7 +2,7 @@ package keeper
 
 import (
 	"fmt"
-	"github.com/okex/exchain/libs/cosmos-sdk/x/supply/internal/blockfee"
+	"github.com/okex/exchain/libs/cosmos-sdk/x/supply/internal/fee"
 
 	"github.com/okex/exchain/libs/tendermint/libs/log"
 
@@ -19,7 +19,7 @@ type Keeper struct {
 	ak        types.AccountKeeper
 	bk        types.BankKeeper
 	permAddrs map[string]types.PermissionsForAddress
-	blockfee.Pool
+	fee.Fee
 }
 
 // NewKeeper creates a new Keeper instance
@@ -36,7 +36,7 @@ func NewKeeper(cdc *codec.Codec, key sdk.StoreKey, ak types.AccountKeeper, bk ty
 		ak:        ak,
 		bk:        bk,
 		permAddrs: permAddrs,
-		Pool:      blockfee.NewCollector(),
+		Fee:       fee.NewCollector(),
 	}
 }
